tests: move FuzzParse to separate file
