+++ pkg/parser/parser_test.go
@@ -19,7 +19,6 @@ package parser
 import (
 	"bytes"
 	"context"
-	"io"
 	"testing"
 
 	"github.com/google/go-cmp/cmp"
@@ -178,14 +177,4 @@ func TestParser(t *testing.T) {
 			}
 		})
 	}
-}
-
-func FuzzParse(f *testing.F) {
-	f.Fuzz(func(t *testing.T, data []byte) {
-		objScheme := runtime.NewScheme()
-		metaScheme := runtime.NewScheme()
-		p := New(metaScheme, objScheme)
-		r := io.NopCloser(bytes.NewReader(data))
-		_, _ = p.Parse(context.Background(), r)
-	})
 }
